init commit of source code for my spring framework course
This contains the initial commit for the project-management application that you will build as part of my spring framework course.
Don't just download the repo and have the code while you take the course!!! I'd like you as the student to actually code along with the videos in your fresh IDE. this code-base is just a reference. This contains the latest and greatest version of the completed project.